feature: add absolute paths
