+++ src/Components/Heading.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 
 const Heading = () => {
   return (
