+++ src/Pages/Home.tsx
@@ -1,19 +1,32 @@
-import Heading from '../Components/Heading'
-import TaskDone from '../Components/TaskDone'
+import Heading from '@/Components/Heading'
+import TaskDone from '@/Components/TaskDone'
 import { useNavigate } from 'react-router-dom'
+import { useState, useMemo } from 'react'
 
-type HomeProps = 
-{
-    doneTask:number;
-    totalTask:number
-}
-const Home = (props:HomeProps) => {
+type Task = { id: number; done: boolean; text: string };
+const Home = () => {
 
+    const [tasks, _] = useState<Task[]>(() => {
+      const save = localStorage.getItem("task");
+      return save ? JSON.parse(save) : [];
+    });
+  
+  
+    const totalTask = tasks.length;
+  
     const navigate = useNavigate();
+      const doneTask = useMemo(
+        () =>
+          tasks.reduce(
+            (tasksCount, currentTask) => tasksCount + (currentTask.done ? 1 : 0),
+            0
+          ),
+        [tasks]
+      );
   return (
     <>
     <Heading />
-    <TaskDone doneCount={props.doneTask} totalCount={props.totalTask} />
+    <TaskDone doneCount={doneTask} totalCount={totalTask} />
     <div className="flex justify-center gap-20">
       <button
         className="text-stone-50 bg-[#8aad32] px-8 py-4 rounded-full cursor-pointer"
